requirements.txt
awscli==1.19.62
boto3==1.17.62
botocore==1.20.62
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
dataclasses-json==0.5.3
docutils==0.15.2
idna==2.10
isodate==0.6.0
jmespath==0.10.0
marshmallow==3.11.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
numpy==1.20.2
oauthlib==3.1.0
pandas==1.2.4
pyasn1==0.4.8
python-dateutil==2.8.1
python-decouple==3.4
python-youtube==0.8.0
pytube==10.7.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.4.2
selenium==3.141.0
six==1.15.0
stringcase==1.2.0
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
yt-videos-list==0.5.5
